Fix AddMembershipModal exports
